+++ datacrunch/volumes/volumes.py
@@ -18,7 +18,7 @@ class Volume:
         is_os_volume: bool,
         created_at: str,
         target: str = None,
-        location: str = Locations.FIN_01,
+        location: str = Locations.FIN_03,
         instance_id: str = None,
         ssh_key_ids: List[str] = [],
         deleted_at: str = None,
@@ -41,7 +41,7 @@ class Volume:
         :type created_at: str
         :param target: target device e.g. vda
         :type target: str, optional
-        :param location: datacenter location, defaults to "FIN-01"
+        :param location: datacenter location, defaults to "FIN-03"
         :type location: str, optional
         :param instance_id: the instance id the volume is attached to, None if detached
         :type instance_id: str
@@ -250,7 +250,7 @@ class VolumesService:
         name: str,
         size: int,
         instance_id: str = None,
-        location: str = Locations.FIN_01,
+        location: str = Locations.FIN_03,
     ) -> Volume:
         """Create new volume
 
@@ -262,7 +262,7 @@ class VolumesService:
         :type size: int
         :param instance_id: Instance id to be attached to, defaults to None
         :type instance_id: str, optional
-        :param location: datacenter location, defaults to "FIN-01"
+        :param location: datacenter location, defaults to "FIN-03"
         :type location: str, optional
         :return: the new volume object
         :rtype: Volume